requirements.txt
fastapi
uvicorn
pandas
prophet
python-dotenv
requests
httpx
